Add check COUNT products before delete Tax Class Product

+++ app/code/core/Mage/Adminhtml/controllers/Tax/Class/ProductController.php
@@ -104,6 +104,16 @@ class Mage_Adminhtml_Tax_Class_ProductController extends Mage_Adminhtml_Controll
             return;
         }
 
+        $productCollection = Mage::getModel('catalog/product')
+            ->getCollection()
+            ->addAtrributeToFilter('tax_class_id', $classId);
+        $productCount = $productCollection->getSize();
+        if ($productCount > 0) {
+            Mage::getSingleton('adminhtml/session')->addError(Mage::helper('tax')->__('You cannot delete this tax class as it is used for %d products.', $productCount));
+            $this->_redirectReferer();
+            return;
+        }
+
         try {
             $classModel->delete();
